src/main/java/krzysiek/springframework/recipespring5/converters/CategoryToCategoryCommand.java
package krzysiek.springframework.recipespring5.converters;

import krzysiek.springframework.recipespring5.commands.CategoryCommand;
import krzysiek.springframework.recipespring5.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

    @Synchronized
    @Nullable
    @Override
    public CategoryCommand convert(Category source) {
        if (source == null) {
            return null;
        }
        final CategoryCommand categoryCommand = new CategoryCommand();

        categoryCommand.setId(source.getId());
        categoryCommand.setDescription(source.getDescription());

        return categoryCommand;
    }
}
